+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
-## [Unreleased]
+## [0.22.0] - 2021-06-04
 
 ### Added
 
@@ -592,7 +592,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Support for internationalizing pages.
 - Both light and dark themes for the site.
 
-[unreleased]: https://github.com/visiosto/visiosto-website/compare/v0.21.0...HEAD
+[unreleased]: https://github.com/visiosto/visiosto-website/compare/v0.22.0...HEAD
+[0.22.0]: https://github.com/visiosto/visiosto-website/compare/v0.21.0...v0.22.0
 [0.21.0]: https://github.com/visiosto/visiosto-website/compare/v0.20.1...v0.21.0
 [0.20.1]: https://github.com/visiosto/visiosto-website/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/visiosto/visiosto-website/compare/v0.19.1...v0.20.0